Extract z.ai provider tests
